# public/openapi.yml
openapi: 3.1.0
servers:
  - url: "https://vulnerabilityhistory.org"
    description: The main source of API calls
  - url: "https://cnd.vulnerabilityhistory.org"
    description: A cached source of data, but generally faster
info:
  title: Vulnerability History Project RESTful API
  description: |
    This read-only API allows detailed access to the VHP data.

    This documentation is a work-in-progress. If you want to contribute, see [our issue on GitHub](https://github.com/VulnerabilityHistoryProject/vulnerability-history/issues/956).
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
security: []
...
paths:
...
  /api/tags:
    get:
      summary: /api/tags
      operationId: listTags
      tags:
        - tags
      description: List all tags in the database
      parameters:
        - name: map
          description: |
            Change the return format from array to dictionary
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      format: int32
                    name:
                      type: string
                    shortname:
                      type: string
                    description:
                      type: string
                    color:
                      type: string
                    icon:
                      type: string
                    short_desc:
                      type: string

              example:
                - id: 373
                  name: 'Chromium subsystem: autofill'
                  shortname: subsystem-chromium-autofill
                  description: |2+
                    What subsystem was this in?

                    Most software project don't make a definitive list of their subsystems, so the subsystems tags are our curators' best guess at what part(s) of the system a vulnerability was in. Often, curators will look at file paths primarily, but may also look at bug reports or commit messages to make this determination.

                    As part of the editorial process, we periodically combine subsystems together. If you see two subsystems that ought to be combined, then do a pull request against the vulnerabilities involved! We welcome that kind of edit.

                  color: "#d9e1ed"
                  icon: cogs
                  short_desc: |2-
                    What subsystem was this in?

                    Most software project don't make a definitive list of their subsystems, so the subsystems tags are our curators' best guess at what part(s) of the system a vulnerability was in. Often, curators will look at f...
                - id: 3
                  name: 'Lesson: Code Refactors'
                  shortname: refactors
                  description: |-
                    The vulnerable code had a lot of refactors in its lifetime.

                    A vulnerability with this tag has had at least 10 separate weeks where at least one commit mentioned the word "refactoring" in the commit message.

                    Refactoring is generally considered to be a good practice to be doing regularly. It indicates regular maintenance and attention. While refactoring doesn't yield new features or fix defects, it greatly reduces the cost of them.
                  color: "#009696"
                  icon: branch
                  short_desc: |-
                    The vulnerable code had a lot of refactors in its lifetime.

                    A vulnerability with this tag has had at least 10 separate weeks where at least one commit mentioned the word "refactoring" in the commit message.
        "400":
          $ref: '#/components/responses/400Malformed'
        "500":
          $ref: '#/components/responses/500Internal'
...
  /api/tags/overlaps/array:
    get:
      summary: /api/tags/overlaps/array
      operationId: Overlaps
      tags:
        - overlaps
      description: |
        Returns the vulnerabilities in common between the tags provided in decending
        order according to upvotes
      parameters:
        - name: overlaps
          description: |
            The tags to find the overlaps of.
          in: query
          required: true
          schema:
            type: array of tag shortnames
            items:
              type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  nodes:
                    $ref: '#/components/schemas/sankeyNodes'
                  links:
                    $ref: '#/components/schemas/sankeyLinks'
              example:
        "400":
          $ref: '#/components/responses/400Malformed'
        "500":
          $ref: '#/components/responses/500Internal'
...
components:
  responses:
    400Malformed: # Can be referenced as '#/components/responses/400Malformed'
      description: Malformed request.
    500Internal: # Can be referenced as '#/components/responses/500Internal'
      description: |
        Internal error. Please report these on our
        [issue tracker](https://github.com/VulnerabilityHistoryProject/vulnerability-history/issues).
      content: {}
  schemas:
...
    sankeyNodes:
      type: array
      items:
        type: object
        properties:
          name:
            type: string
    sankeyLinks:
      type: array
      items:
        type: object
        properties:
          source:
            type: string
          target:
            type: string
          value: 
            type: integer
...